feat(internal): Allow for specifying optional measurement units
The value is then converted to mebibytes by dividing
the integer by 1024*1024. If something like ~3.2Mi
are specified, the value will be truncated down to 3Mi.

+++ internal/cli/kraft/cloud/deploy/deployer_image_name.go
@@ -97,7 +97,7 @@ func (deployer *deployerImageName) Deploy(ctx context.Context, opts *DeployOptio
 					Features:               opts.Features,
 					Domain:                 opts.Domain,
 					Image:                  deployer.imageName,
-					Memory:                 uint(opts.Memory),
+					Memory:                 opts.Memory,
 					Metro:                  opts.Metro,
 					Name:                   opts.Name,
 					Ports:                  opts.Ports,
